Simplify Policy evaluation
